database: Refactoring enum-checking model.
